correctly initialize a fresh database
The current code incorrectly handles the case when the database does not exist. It fails to create the database schema. It only creates the schema when the "--remove" flag is passed, but this shouldn't be necessary when the database does not exist, as there is nothing to remove.

This PR addresses the issue.